requirements.txt
# Core Deep Learning Framework
# Install PyTorch with CUDA support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Computer Vision and Object Detection
ultralytics>=8.0.200
timm>=0.9.7
opencv-python>=4.8.0

# Data Processing and Augmentation
albumentations>=1.3.1
numpy>=1.24.0
Pillow>=9.5.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and Logging
tqdm>=4.65.0
tensorboard>=2.13.0
loguru>=0.7.0

# Model Export and Optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Machine Learning Utilities
scikit-learn>=1.3.0
scipy>=1.10.0

# COCO Dataset Tools
pycocotools>=2.0.6

# Configuration and File I/O
PyYAML>=6.0
omegaconf>=2.3.0